+++ api-conversion/src/main/java/io/strimzi/kafka/api/conversion/cli/AbstractCommand.java
@@ -8,6 +8,16 @@ import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.dsl.MixedOperation;
 import io.strimzi.api.annotations.ApiVersion;
 import io.strimzi.api.kafka.Crds;
+import io.strimzi.api.kafka.model.Kafka;
+import io.strimzi.api.kafka.model.KafkaBridge;
+import io.strimzi.api.kafka.model.KafkaConnect;
+import io.strimzi.api.kafka.model.KafkaConnectS2I;
+import io.strimzi.api.kafka.model.KafkaConnector;
+import io.strimzi.api.kafka.model.KafkaMirrorMaker;
+import io.strimzi.api.kafka.model.KafkaMirrorMaker2;
+import io.strimzi.api.kafka.model.KafkaRebalance;
+import io.strimzi.api.kafka.model.KafkaTopic;
+import io.strimzi.api.kafka.model.KafkaUser;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -16,7 +26,6 @@ import picocli.CommandLine;
 import java.util.Map;
 import java.util.Set;
 import java.util.function.BiFunction;
-import java.util.function.Function;
 
 public abstract class AbstractCommand implements Runnable {
     protected Logger log = LogManager.getLogger(getClass().getName());
@@ -37,24 +46,24 @@ public abstract class AbstractCommand implements Runnable {
             "KafkaRebalance"
     );
 
+    // Old API versions used by given resource => used to obtain the unconverted resource
+    @SuppressWarnings("deprecation")
+    protected final static Map<String, String> OLD_API_VERSIONS = Map.of(
+            "Kafka", Kafka.V1BETA1,
+            "KafkaConnect", KafkaConnect.V1BETA1,
+            "KafkaConnectS2I", KafkaConnectS2I.V1BETA1,
+            "KafkaMirrorMaker", KafkaMirrorMaker.V1BETA1,
+            "KafkaBridge", KafkaBridge.V1ALPHA1,
+            "KafkaMirrorMaker2", KafkaMirrorMaker2.V1ALPHA1,
+            "KafkaTopic", KafkaTopic.V1BETA1,
+            "KafkaUser", KafkaUser.V1BETA1,
+            "KafkaConnector", KafkaConnector.V1ALPHA1,
+            "KafkaRebalance", KafkaRebalance.V1ALPHA1
+    );
+
     // Versioned operations are used to write the converted resource using the target API
     @SuppressWarnings({"rawtypes"})
     protected final static Map<String, BiFunction<KubernetesClient, String, MixedOperation>> VERSIONED_OPERATIONS = Map.of(
-            "Kafka", Crds::kafkaOperation,
-            "KafkaConnect", Crds::kafkaConnectOperation,
-            "KafkaConnectS2I", Crds::kafkaConnectS2iOperation,
-            "KafkaMirrorMaker", Crds::mirrorMakerOperation,
-            "KafkaBridge", Crds::kafkaBridgeOperation,
-            "KafkaMirrorMaker2", Crds::kafkaMirrorMaker2Operation,
-            "KafkaTopic", Crds::topicOperation,
-            "KafkaUser", Crds::kafkaUserOperation,
-            "KafkaConnector", Crds::kafkaConnectorOperation,
-            "KafkaRebalance", Crds::kafkaRebalanceOperation
-    );
-
-    // The operation with the default versions
-    @SuppressWarnings({"rawtypes"})
-    protected final static Map<String, Function<KubernetesClient, MixedOperation>> DEFAULT_OPERATIONS = Map.of(
             "Kafka", Crds::kafkaOperation,
             "KafkaConnect", Crds::kafkaConnectOperation,
             "KafkaConnectS2I", Crds::kafkaConnectS2iOperation,