+++ api/src/main/java/io/strimzi/api/kafka/Crds.java
@@ -69,7 +69,6 @@ public class Crds {
         String version = null;
         if (cls.equals(Kafka.class)) {
             version = Kafka.CONSUMED_VERSION;
-            //version = KafkaConnect.VERSIONS.get(0);
         } else if (cls.equals(KafkaConnect.class)) {
             version = KafkaConnect.CONSUMED_VERSION;
         } else if (cls.equals(KafkaConnectS2I.class)) {