+++ api/src/test/java/io/strimzi/api/kafka/model/AbstractCrdIT.java
@@ -19,19 +19,12 @@ import static org.hamcrest.CoreMatchers.containsStringIgnoringCase;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assumptions.assumeTrue;
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 public abstract class AbstractCrdIT {
 
     protected KubeClusterResource cluster = KubeClusterResource.getInstance();
-
-    protected void assumeKube1_11Plus() {
-        VersionInfo version = new DefaultKubernetesClient().getVersion();
-        assumeTrue("1".equals(version.getMajor())
-                && Integer.parseInt(version.getMinor().split("\\D")[0]) >= 11);
-    }
 
     protected void assumeKube1_16Plus() {
         VersionInfo version = new DefaultKubernetesClient().getVersion();
@@ -49,7 +42,7 @@ public abstract class AbstractCrdIT {
     protected <T extends CustomResource> void createDelete(Class<T> resourceClass, String resource) {
         T model = loadResource(resourceClass, resource);
         String modelStr = TestUtils.toYamlString(model);
-        assertDoesNotThrow(() -> createDelete(modelStr), "Create delete failed after first round-trip -- maybe a problem with a defaulted value?\nApplied string: " + modelStr);
+        createDelete(modelStr);
     }
 
     private void createDelete(String ssStr) {
@@ -117,7 +110,8 @@ public abstract class AbstractCrdIT {
         for (String requiredProperty: requiredProperties) {
             assertThat("Could not find" + requiredProperty + " in message: " + message, message, anyOf(
                     containsStringIgnoringCase(requiredProperty + " in body is required"),
-                    containsStringIgnoringCase(requiredProperty + ": Required value")
+                    containsStringIgnoringCase(requiredProperty + ": Required value"),
+                    containsStringIgnoringCase("missing required field \"" + requiredProperty + "\"")
             ));
         }
     }
@@ -127,8 +121,16 @@ public abstract class AbstractCrdIT {
             JsonNode json = (JsonNode) crd;
             if (json != null
                     && json.hasNonNull("status")
-                    && json.get("status").hasNonNull("conditions")) {
-                return true;
+                    && json.get("status").hasNonNull("conditions")
+                    && json.get("status").get("conditions").isArray()) {
+                for (JsonNode condition : json.get("status").get("conditions")) {
+                    if ("Established".equals(condition.get("type").asText())
+                            && "True".equals(condition.get("status").asText()))   {
+                        return true;
+                    }
+                }
+
+                return false;
             }
 
             return false;