+++ systemtest/src/main/java/io/strimzi/systemtest/resources/crd/KafkaConnectResource.java
@@ -33,7 +33,7 @@ public class KafkaConnectResource {
     public static final String PATH_TO_KAFKA_CONNECT_CONFIG = TestUtils.USER_PATH + "/../packaging/examples/connect/kafka-connect.yaml";
 
     public static MixedOperation<KafkaConnect, KafkaConnectList, Resource<KafkaConnect>> kafkaConnectClient() {
-        return Crds.kafkaConnectV1Beta2Operation(ResourceManager.kubeClient().getClient());
+        return Crds.kafkaConnectOperation(ResourceManager.kubeClient().getClient());
     }
 
     public static KafkaConnectBuilder kafkaConnect(String name, int kafkaConnectReplicas) {