+++ systemtest/src/test/java/io/strimzi/systemtest/tracing/TracingST.java
@@ -30,6 +30,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import io.strimzi.systemtest.resources.KubernetesResource;
@@ -69,6 +70,7 @@ import static org.junit.jupiter.api.Assumptions.assumeFalse;
 @Tag(TRACING)
 @Tag(INTERNAL_CLIENTS_USED)
 @ExtendWith(VertxExtension.class)
+@Disabled
 public class TracingST extends AbstractST {
 
     private static final String NAMESPACE = "tracing-cluster-test";